chore: a lot of stuff
Changed the way I represent the actual valid instructions, which needs further refinement, but works for now.

Also broke out the Block construct to it's own file, to clean up the mod.rs, which works like a scratch pad of stuff for now.

Some fixes for showing unknown instructions and better handling of memory dumps.