fix: use interior mutability for the memory
To be able to share the memory between the
different parts of the c64, I needed to use
Rc to share the ownership, and RefCell to
be able to mutate it from different objects.

This is my first foray into doing this, so I hope
I did the right things here, but I guess there
most certainly is other ways to achieve this?